requirements.txt
streamlit>=1.30.0
ollama>=0.3.0
python-dotenv>=1.0.0
pandas>=2.1.0
plotly>=5.18.0
PyYAML>=6.0.1
reportlab>=4.0.0
requests>=2.31.0
